add precise query

Match-id-b856629b4fe6d184a1ff2f3b9925efacfa1d91b9

+++ services/crypto_manager/src/crypto.rs
@@ -333,7 +333,7 @@ impl Crypto {
         };
 
         // take the AEAD from cipher.
-        if cipher.len() < AEAD_SIZE {
+        if cipher.len() < AEAD_SIZE as usize {
             return Err(ErrCode::Failed);
         }
         let cipher_without_aead_size = cipher.len() - AEAD_SIZE as usize;
